requirements.txt
Pillow
keyboard
pyscreenshot
moviepy
numpy
matplotlib
progress